+++ client/my-app/src/App.js
@@ -15,7 +15,6 @@ function App() {
     <div className="App">
       <SortData />
       <PredictionMatrix />
-      <EstimatAverage />
       <DatePickerComponent />
       <ChartData />
     </div>
