+++ server/api/apiGetDollarRate.js
