+++ server/app.js
@@ -3,11 +3,11 @@ require('dotenv').config();
 const express = require("express");
 const app = express();
 
-const getData = require('./db/findAllData.js')
+const getData = require('./routers/findAllData.js')
 const insertToDb = require('./db/iInsertToDb.js')
-const getAvaregeBymonth = require('./db/findByMonth')
-const apiedForecast = require('./services/apiEstimatedForecast')
-const apiEstimatAllData  = require('./db/allForecast')
+const getAvaregeBymonth = require('./routers/findByMonth')
+const apiedForecast = require('./routers/apiEstimatedForecast')
+const apiEstimatAllData  = require('./routers/allForecast')
 const PORT = process.env.PORT
 
 app.use(express.urlencoded({ extended: true }));
@@ -16,7 +16,7 @@ app.use(express.json());
 
 app.use('/data', getData);
 app.use('/dollarValue', getAvaregeBymonth);
-app.use('/Estimat', apiedForecast );
+// app.use('/Estimat', apiedForecast );
 app.use('/EstimatAllData', apiEstimatAllData );
